+++ rust/dbn/src/lib.rs
@@ -1,10 +1,6 @@
 //! A crate for reading DBN and legacy DBZ files and converting them to other
 //! [`Encoding`](enums::Encoding)s.
 
-// #[cfg_attr(not(feature = "python"), macro_use)]
-// #[cfg(not(feature = "python"))]
-// extern crate dbn_macros;
-
 #[deny(missing_docs)]
 #[deny(rustdoc::broken_intra_doc_links)]
 #[deny(clippy::missing_errors_doc)]